+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Feature: Added `std:sort`, `std:cmp:num` and `std:cmp:str`.
 * Feature: Added `std:copy` for shallow copies.
 * Feature: Added `std:unshift`.
+* Change: Made calling `$none` an invalid action that causes a panic.
 * Bugfix: Correctly handle modification of lists or maps that are being
 iterated over. They cause a WLambda panic now with a backtrace instead of
 panicing on borrow\_mut() in the host application.